move serial to a context
